feat(broker): deploy larger workflows
* increase SBE varDataEncoding length
* configure max message size in broker and gateway
* setting max message size in Dispatcher, Transport and Atomix
* remove buffer size configuration in favor of max message size (configure the buffer size depending on the max message size needs too much internal knowledge)
* distribute deployment with actual length
* limit job batch activation to max message size

+++ protocol-impl/src/main/java/io/zeebe/protocol/impl/record/RecordMetadata.java
@@ -109,7 +109,7 @@ public class RecordMetadata implements BufferWriter, BufferReader {
     if (rejectionReason.capacity() > 0) {
       encoder.putRejectionReason(rejectionReason, 0, rejectionReason.capacity());
     } else {
-      buffer.putShort(offset, (short) 0);
+      buffer.putInt(offset, 0);
     }
   }
